Refactor server process management by separating platform-specific logic into distinct files for Windows and Linux.
